feat: Use redis as a queue
This uses redis to submit jobs to a queue when the initial
generate command is received. A new process - worker - runs
to receive jobs from the generate queue and it will then run
./scripts/generate.sh or similar script and upload to AWS S3.

Once the worker binary is done, it submits the a job to the
results queue. This queue is polled by the bot, who then
sends a GiHub comment with the result.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>