bot: Remove comment when pull request gets opened
We normally post a comment when a PR is first posted, instructing
users on how to make use of the bot. While we're in a period of
migration and getting things working on the main taxonomy repo, turn
this off so it doesn't draw attention from people that shouldn't be
bothered by it yet.

Signed-off-by: Russell Bryant <rbryant@redhat.com>