Convert GitHub bot to go
We have a worker in Go and the GitHub interface side in typescript. I
wanted to see what a Go version of the GitHub side would look like.

It's definitely more verbose, but it's also nice to stick to one
language.

This also includes a new test mode for the worker. It makes it skip
all the real work and immediately queue up a result for a job. That
makes it a lot easier to test an end-to-end-workflow.

Signed-off-by: Russell Bryant <rbryant@redhat.com>