worker: Add retry loop around git fetch
We have the same retry loop around a checkout just after this and I
saw the error happen on this fetch in prod. I don't know what causes
this error, but a retry seems to work.

If we keep hitting weird behavior like this, moving off of `go-git` to
something that uses real `git` underneath seems like a good idea.

instruct-lab-bot-worker[54466]: 2024-04-03T21:59:35.026Z        ERROR
    cmd/generate.go:303        Could not fetch from origin: reference has
    changed concurrently        {"job": "25", "pr_number": "620",
    "work_dir": "/home/fedora/instruct-lab-bot", "origin": "origin"}

    github.com/instruct-lab/instruct-lab-bot/worker/cmd.processJob
        /home/fedora/instruct-lab-bot/bot-repo/worker/cmd/generate.go:303
    github.com/instruct-lab/instruct-lab-bot/worker/cmd.glob..func1.3
        /home/fedora/instruct-lab-bot/bot-repo/worker/cmd/generate.go:129
```

Signed-off-by: Russell Bryant <rbryant@redhat.com>