Restore bot introduction on new pull requests
The original bot commented on PRs as soon as they came in to introduce
the bot. We turned this off while doing more development and testing
against the main taxonomy repo, but I think we're ready to turn this
back on and let everyone use it who wants to.

The introduction lists the commands, as well as the required label if
one has been configured.

Closes #82

Signed-off-by: Russell Bryant <rbryant@redhat.com>