Add redis pooling and worker refactor
- Added a redis pool to resolve #165
- Some refactoring to create an instance of a new worker
struct for each new job.
- Added job ID to results for better debugability and tracking.
- Added a precheck error reporting. One thing of note is if no yaml
files were submitted, chat would run without errors resulting in
an empty report. This checks if no yaml files are in the diff
and exits the runPrecheck() method early to avoid unnecessary
processing since precheck could be a default action on PRs. It
also avoids an empty report.

Signed-off-by: Brent Salisbury <bsalisbu@redhat.com>