Only try to get remote model name for precheck
In prod, the only case we want to try to get a remote model name is
under `precheck`, where we use `ilab chat` against a remote endpoint
URL.

The code for endpoint URLs under `generate-local` is commented out,
but that's because we only have a single `--endpoint-url` option. We
need to split it between `--chat-endpoint-url` and
`--generate-endpoint-url` or something like that, so we can say
specifically which endpoint to use with each command.

In the meantime though, we need to make sure we don't hit the
endpoint when we're actually running `endpoint-local`.

Signed-off-by: Russell Bryant <rbryant@redhat.com>