Remove explicit enable command, only maintainer teams member will be able to run commands.
Bot will post a welcome message when required labels
are set on the PR (currently set to skill and knowledge).

Our bot command name is not changed to @instructlab-bot
from @instruct-lab-bot, due to renaming at the instruct lab
organization level. But we do support the legacy bot username
but post a deprecated warning on the PR.

Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>