Use github status for checking bot readiness.
Checks are not immidiately available to fetch, but
user can add multiple labels pretty fast, that can
bombard the PR with multiple bot welcome message.
Status can be immidiately retried after posting,
and that's helpful in this scenario.

Also added "help" command, if user wants to dump the
welcome message. We can use this command to post more
helpful text to the PR.

Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>